Refactor: brainfart: `crush_n` should never have lived in `preprocess`. It has its own module now
